Move props to own package
